+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/privacy/methods/eea/EeaSendRawTransaction.java
@@ -14,13 +14,8 @@
  */
 package org.hyperledger.besu.ethereum.api.jsonrpc.internal.privacy.methods.eea;
 
-import static org.apache.logging.log4j.LogManager.getLogger;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.JsonRpcEnclaveErrorConverter.convertEnclaveInvalidReason;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.JsonRpcErrorConverter.convertTransactionInvalidReason;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.DECODE_ERROR;
-import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.PRIVATE_FROM_DOES_NOT_MATCH_ENCLAVE_PUBLIC_KEY;
-import static org.hyperledger.besu.ethereum.privacy.PrivacyGroupUtil.findOffchainPrivacyGroup;
-
+import org.apache.logging.log4j.Logger;
+import org.apache.tuweni.bytes.Bytes;
 import org.hyperledger.besu.enclave.types.PrivacyGroup;
 import org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequestContext;
@@ -42,8 +37,12 @@ import org.hyperledger.besu.ethereum.transaction.TransactionInvalidReason;
 
 import java.util.Optional;
 
-import org.apache.logging.log4j.Logger;
-import org.apache.tuweni.bytes.Bytes;
+import static org.apache.logging.log4j.LogManager.getLogger;
+import static org.hyperledger.besu.ethereum.api.jsonrpc.JsonRpcEnclaveErrorConverter.convertEnclaveInvalidReason;
+import static org.hyperledger.besu.ethereum.api.jsonrpc.JsonRpcErrorConverter.convertTransactionInvalidReason;
+import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.DECODE_ERROR;
+import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.PRIVATE_FROM_DOES_NOT_MATCH_ENCLAVE_PUBLIC_KEY;
+import static org.hyperledger.besu.ethereum.privacy.PrivacyGroupUtil.findOffchainPrivacyGroup;
 
 public class EeaSendRawTransaction implements JsonRpcMethod {
 
@@ -85,8 +84,7 @@ public class EeaSendRawTransaction implements JsonRpcMethod {
       final Optional<Bytes> maybePrivacyGroupId = privateTransaction.getPrivacyGroupId();
 
       final Optional<PrivacyGroup> maybePrivacyGroup =
-          findPrivacyGroup(
-              privacyController, maybePrivacyGroupId, enclavePublicKey, privateTransaction);
+          findPrivacyGroup(maybePrivacyGroupId, enclavePublicKey, privateTransaction);
 
       final ValidationResult<TransactionInvalidReason> validationResult =
           privacyController.validatePrivateTransaction(privateTransaction, enclavePublicKey);
@@ -96,15 +94,19 @@ public class EeaSendRawTransaction implements JsonRpcMethod {
             id, convertTransactionInvalidReason(validationResult.getInvalidReason()));
       }
 
-      final JsonRpcResponse ret =
-          createPMTAndAddToTxPool(
+      final Transaction privacyMarkerTransaction =
+          createPMT(
               id,
               privateTransaction,
               maybePrivacyGroup,
               maybePrivacyGroupId,
-              enclavePublicKey,
-              Address.DEFAULT_PRIVACY);
-      return ret;
+              enclavePublicKey);
+
+      return transactionPool
+              .addLocalTransaction(privacyMarkerTransaction)
+              .either(
+                      () -> new JsonRpcSuccessResponse(id, privacyMarkerTransaction.getHash().toString()),
+                      errorReason -> getJsonRpcErrorResponse(id, errorReason));
 
     } catch (final JsonRpcErrorResponseException e) {
       return new JsonRpcErrorResponse(id, e.getJsonRpcError());
@@ -117,7 +119,6 @@ public class EeaSendRawTransaction implements JsonRpcMethod {
   }
 
   Optional<PrivacyGroup> findPrivacyGroup(
-      final PrivacyController privacyController,
       final Optional<Bytes> maybePrivacyGroupId,
       final String enclavePublicKey,
       final PrivateTransaction privateTransaction) {
@@ -126,27 +127,20 @@ public class EeaSendRawTransaction implements JsonRpcMethod {
     return maybePrivacyGroup;
   }
 
-  JsonRpcResponse createPMTAndAddToTxPool(
-      final Object id,
-      final PrivateTransaction privateTransaction,
-      final Optional<PrivacyGroup> maybePrivacyGroup,
-      final Optional<Bytes> maybePrivacyGroupId,
-      final String enclavePublicKey,
-      final Address privacyPrecompileAddress) {
+  Transaction createPMT(
+          final Object id,
+          final PrivateTransaction privateTransaction,
+          final Optional<PrivacyGroup> maybePrivacyGroup,
+          final Optional<Bytes> maybePrivacyGroupId,
+          final String enclavePublicKey) {
     final String privateTransactionLookupId =
         privacyController.sendTransaction(privateTransaction, enclavePublicKey, maybePrivacyGroup);
-    final Transaction privacyMarkerTransaction =
-        privacyController.createPrivacyMarkerTransaction(
-            privateTransactionLookupId, privateTransaction, privacyPrecompileAddress);
-    return transactionPool
-        .addLocalTransaction(privacyMarkerTransaction)
-        .either(
-            () -> new JsonRpcSuccessResponse(id, privacyMarkerTransaction.getHash().toString()),
-            errorReason -> getJsonRpcErrorResponse(id, errorReason));
+    return privacyController.createPrivacyMarkerTransaction(
+            privateTransactionLookupId, privateTransaction, Address.DEFAULT_PRIVACY);
   }
 
   JsonRpcErrorResponse getJsonRpcErrorResponse(
-      final Object id, final TransactionInvalidReason errorReason) {
+          final Object id, final TransactionInvalidReason errorReason) {
     if (errorReason.equals(TransactionInvalidReason.INTRINSIC_GAS_EXCEEDS_GAS_LIMIT)) {
       return new JsonRpcErrorResponse(id, JsonRpcError.PMT_FAILED_INTRINSIC_GAS_EXCEEDS_LIMIT);
     }
